fix: don't mix dates
We should be querying for the year 2022, according to the text we're
generating. Also, we're now using the year variable to have it all in
just one place (and when we decide to run this for 2023, we just need
to update one variable).

In the future we might want to get the year as a parameter.